Work on supporting vector fonts

+++ modules/mod_text/mod_text.c
@@ -261,9 +261,9 @@ int modtext_text_height(INSTANCE *my, int *params) {
 
 int modtext_text_width(INSTANCE *my, int *params) {
     const char *str = string_get(params[1]);
-    int result = gr_text_width(params[0], (const unsigned char *)str);
+    uint32_t result = gr_text_width(params[0], (const unsigned char *)str);
     string_discard(params[1]);
-    return result;
+    return (int)result;
 }
 
 /* --------------------------------------------------------------------------- */
